requirements.txt
Django==3.1.14
djangorestframework==3.11.2
Markdown==3.1
psycopg2==2.8.5
pytz==2019.3
django-taggit==1.2.0
django-polymorphic==3.1.0
coverage==5.1
django-crispy-forms==1.9.0
Pillow==9.0.1
django-cleanup==4.0.0
reportlab==3.5.55
djangorestframework-datatables==0.5.2
mozilla_django_oidc==2.0.0
gunicorn==20.0.4
django-select2==7.6.1
tmdbv3api==1.7.6
django-cors-headers==3.8.0
#djangorestframework-simplejwt==4.8.0
openlibrary-client==0.0.30
django_redis==5.2.0